# src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mithra Direct App",
    "description": "API documentation for the subscription services",
    "contact": {
      "name": "Swamy Kunta",
      "url": "https://github.com/s713278/ECommerceApp/tree/master",
      "email": "swamy.kunta@gmail.com"
    },
    "license": {
      "name": "License",
      "url": "/"
    },
    "version": "v1"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Env"
    },
    {
      "url": "http://172.105.33.153:8080/api",
      "description": "Dev Env"
    }
  ],
  "paths": {
...
    "/v1/auth/verify-otp": {
      "post": {
        "tags": [
          "1. User Authentication"
        ],
        "summary": "Verify OTP",
        "operationId": "verifyMobileOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPVerificationRequest"
              },
              "examples": {
                "Example": {
                  "summary": "Example to OTP Verify",
                  "value": {
                    "mobile_number": 9912149089,
                    "country_code": "+91",
                    "otp": "9090"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "WithAddressAndVendor": {
                    "summary": "Response with address and vendor",
                    "value": {
                      "timestamp": "2024-12-09T04:21:22.8275021",
                      "success": true,
                      "status": 200,
                      "data": {
                        "address": {
                          "country": "India",
                          "zipCode": "500001",
                          "city": "Hyderabad",
                          "state": "Telangana"
                        },
                        "user_id": 4902,
                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI0OTAyIiwicm9sZXMiOiJbVVNFUl0iLCJpc3MiOiJTdWJOUmVuZXdhbHMiLCJleHAiOjE3MzM3MDAwODIsImlhdCI6MTczMzY5ODI4Mn0.noD8pa5h59ApFIdasIN1WX_4QD3ooV8PEdFwO7AtCgE",
                        "mobile_verified": true,
                        "email_verified": false,
                        "refresh_token": "6e9cadb0-d4df-42a2-af31-3b4b017af852",
                        "default_vendor_id": 1
                      }
                    }
                  },
                  "WithAddressAndWithoutVendor": {
                    "summary": "Response with address and without vendor",
                    "value": {
                      "timestamp": "2024-12-09T04:21:22.8275021",
                      "success": true,
                      "status": 200,
                      "data": {
                        "address": {
                          "country": "India",
                          "zipCode": "500001",
                          "city": "Hyderabad",
                          "state": "Telangana"
                        },
                        "user_id": 4902,
                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI0OTAyIiwicm9sZXMiOiJbVVNFUl0iLCJpc3MiOiJTdWJOUmVuZXdhbHMiLCJleHAiOjE3MzM3MDAwODIsImlhdCI6MTczMzY5ODI4Mn0.noD8pa5h59ApFIdasIN1WX_4QD3ooV8PEdFwO7AtCgE",
                        "mobile_verified": true,
                        "email_verified": false,
                        "refresh_token": "6e9cadb0-d4df-42a2-af31-3b4b017af852",
                        "default_vendor_id": null
                      }
                    }
                  },
                  "WithOutAddressAndWithVendor": {
                    "summary": "Response without address and with vendor",
                    "value": {
                      "timestamp": "2024-12-09T04:21:22.8275021",
                      "success": true,
                      "status": 200,
                      "data": {
                        "address": null,
                        "user_id": 4902,
                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI0OTAyIiwicm9sZXMiOiJbVVNFUl0iLCJpc3MiOiJTdWJOUmVuZXdhbHMiLCJleHAiOjE3MzM3MDAwODIsImlhdCI6MTczMzY5ODI4Mn0.noD8pa5h59ApFIdasIN1WX_4QD3ooV8PEdFwO7AtCgE",
                        "mobile_verified": true,
                        "email_verified": false,
                        "refresh_token": "6e9cadb0-d4df-42a2-af31-3b4b017af852",
                        "default_vendor_id": 1
                      }
                    }
                  },
                  "WithOutAddressAndWithoutVendor": {
                    "summary": "Response without address and vendor",
                    "value": {
                      "timestamp": "2024-12-09T04:21:22.8275021",
                      "success": true,
                      "status": 200,
                      "data": {
                        "address": null,
                        "user_id": 4902,
                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI0OTAyIiwicm9sZXMiOiJbVVNFUl0iLCJpc3MiOiJTdWJOUmVuZXdhbHMiLCJleHAiOjE3MzM3MDAwODIsImlhdCI6MTczMzY5ODI4Mn0.noD8pa5h59ApFIdasIN1WX_4QD3ooV8PEdFwO7AtCgE",
                        "mobile_verified": true,
                        "email_verified": false,
                        "refresh_token": "6e9cadb0-d4df-42a2-af31-3b4b017af852",
                        "default_vendor_id": null
                      }
                    }
                  }
                }
              }
            }
          },
          "401":{
            "description": "Authentication Failed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }, "examples": {
                  "Success": {
                    "summary": "Response for expired/invalid OTP",
                    "value": {
                      "success": false,
                      "status": 401,
                      "timestamp": "2024-12-09T04:55:17.9766051",
                      "failure_reason": "OTP is already used or expired,Please request fow new one",
                      "user_message": "User authentication failed!!",
                      "api_details": []
                    }
                  }
                }
              }
            }
        }
        }
      }
    },
    "/v1/auth/request-otp": {
      "post": {
        "tags": [
          "1. User Authentication"
        ],
        "summary": "Request OTP",
        "operationId": "requestOTP",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPRequest"
              },
              "examples": {
                "User": {
                  "summary": "Example for a User",
                  "value": {
                    "country_code": "+91",
                    "mobile_number": 9912149089,
                    "reg_platform": "android",
                    "user_type": "USER"
                  }
                },
                "Vendor": {
                  "summary": "Example for a Vendor",
                  "value": {
                    "country_code": "+91",
                    "mobile_number": 9912149089,
                    "reg_platform": "android",
                    "user_type": "VENDOR"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "OTPSent": {
                    "summary": "Success Response",
                    "value": {
                      "timestamp": "2024-11-24T08:56:57.8450894",
                      "success": true,
                      "status": 200,
                      "data": "OTP Sent successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "BadMobileNumber": {
                    "summary": "Response for bad mobile number",
                    "value": {
                      "success": false,
                      "status": 400,
                      "timestamp": "2024-12-09T04:53:04.8421849",
                      "failure_reason": "Invalid mobile number and should be 10 digits.",
                      "user_message": "Please check the request and try again!!",
                      "api_details": []
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "1. User Authentication"
        ],
        "summary": "Refresh Token",
        "operationId": "refreshAccessToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/profile": {
      "get": {
        "tags": [
          "1. User Authentication"
        ],
        "summary": "Authenticated Profile Info",
        "operationId": "getAuthenticatedUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "APIResponseObject": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "type": "object",
            "description": "Data"
          }
        }
      },
...
      "OTPVerificationRequest": {
        "required": [
          "country_code",
          "otp"
        ],
        "type": "object",
        "properties": {
          "country_code": {
            "type": "string",
            "description": "Country Code",
            "example": "+91"
          },
          "mobile_number": {
            "type": "integer",
            "format": "int64"
          },
          "otp": {
            "type": "string"
          }
        }
      },
      "OTPRequest": {
        "required": [
          "country_code",
          "mobile_number",
          "reg_platform",
          "user_type"
        ],
        "type": "object",
        "properties": {
          "country_code": {
            "type": "string",
            "description": "Country Code",
            "example": "+91"
          },
          "mobile_number": {
            "type": "integer",
            "format": "int64"
          },
          "reg_platform": {
            "type": "string",
            "description": "source",
            "example": "android"
          },
          "user_type": {
            "type": "string",
            "description": "type",
            "enum": [
              "USER",
              "VENDOR"
            ]
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "E-Commerce Application": {
        "type": "http",
        "in": "header",
        "scheme": "bearer"
      }
    }
  }
}